pyproject.toml
[project]
name = "ai-rewriter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.0",
    "google-genai>=1.63.0",
    "guardrails-ai>=0.8.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
